pyproject.toml
[project]
name = "Intelligence_Server"
version = "3.16.0"
description = "KonaAI Intelligence Server"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "anthropic==0.72.0",
    "azure-identity==1.25.1",
    "azure-keyvault-keys==4.11.0",
    "azure-keyvault-secrets==4.10.0",
    "bcrypt==5.0.0",
    "celery==5.5.3",
    "cohere==5.20.0",
    "dask[complete,dataframe,distributed]==2025.10.0",
    "faiss-cpu==1.12.0",
    "humanize==4.14.0",
    "imbalanced-learn==0.14.0",
    "langchain-community==0.4.1",
    "langchain-google-genai==3.0.0",
    "langchain-openai==1.0.1",
    "lightgbm==4.6.0",
    "mlflow==3.5.1",
    "nicegui==3.2.0",
    "optuna==4.5.0",
    "pandas==2.3.3",
    "pika==1.3.2",
    "psutil==7.0.0",
    "pyarrow>=21.0.0",
    "pydantic-settings==2.11.0",
    "pydantic[email]==2.12.3",
    "pyinstaller==6.16.0",
    "pyodbc==5.3.0",
    "python-jose[cryptography]==3.5.0",
    "sentence-transformers==5.1.2",
    "shap==0.49.1",
    "sqlglot==27.29.0",
    "validators==0.35.0",
    "xgboost==3.1.1",
    "xlsxwriter==3.2.9",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }

[dependency-groups]
dev = [
    "faker>=37.6.0",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "line-profiler>=5.0.0",
    "openpyxl>=3.1.5",
    "pre-commit>=4.3.0",
    "pretty-errors>=1.2.25",
    "pycountry>=24.6.1",
    "pyinstaller>=6.15.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-html>=4.1.1",
    "pytest-sugar>=1.1.1",
    "viztracer>=1.0.4",
]

[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py"]
testpaths = ["tests/"]
addopts = "-s -rf --cov=code --cov-report=html:./build/coverage --html=./build/pytest_report.html --self-contained-html"
log_cli = false
log_cli_level = 1
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
filterwarnings = [
    "ignore::UserWarning"
]
markers = [
    "slow: mark slow tests",
]

[tool.ruff.lint.isort]
order-by-type = true
force-sort-within-sections = true
force-wrap-aliases = true
combine-as-imports = true
lines-after-imports = 1
relative-imports-order = "closest-to-furthest"

[tool.uv]
package = false

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["code"]
